add unit test for convertor
