--- .env
@@ -1,2 +1,0 @@
-API_KEY=c6978d7a29f5d34b0cb9597f322528d6
-CONVERTER_URL=https://api2.online-convert.com/jobs
